third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFilterPrimitiveStandardAttributes-dom-result-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'result' attribute of the SVGFilterPrimitiveStandardAttributes object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.getAttribute('result') is "res1"
PASS gradientElement.getAttribute('result') is "res2"
PASS successfullyParsed is true

TEST COMPLETE
